Quite a major update: -added hardware double click support -in 'fast mode', added movement notifications
Doing so required a few interface changes, sorry for that.

renamed setChipInDynamicMode to resetChip, since that's actually what it is (after finding out that this resets the configured registers..)

Removed the bWakeChipFirst parameter from setAutoSleep, getDeviceType, getFirmwareVersion, sleep.
Rationale: it's every slow and resets the chip state. That can be restored but that makes it even slower.
Typically, these methods will be used either directly after a poweron, or as a result of a touch screen event anyhow and in both of these situations a reset is not needed.

added setMovementInterval, so that the time which should have passed before a new gesture-movement is reported can be set